pyproject.toml
[project]
name = "zoomac"
version = "0.1.0"
description = "Hybrid AI agent with intelligent memory"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sentence-transformers>=2.0",
    "numpy",
    "uvloop>=0.21; sys_platform != 'win32'",
    "pyyaml>=6.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
zoomac = "zoomac.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--basetemp=runner_check/pytest -p no:cacheprovider"
